SE-45 Remove redundant responsibility from swashApiHelper
Remove some functionalities from api helper
Changed communityHelper.ts to userHelper.ts
Changed their occurrences
