+++ src/core/stream.ts
@@ -1,13 +1,13 @@
 import { Message } from '../types/message.type';
 import { Stream } from '../types/stream.type';
 
-import { communityHelper } from './communityHelper';
 import { databaseHelper } from './databaseHelper';
+import { userHelper } from './userHelper';
 
 // Create the client and give the API key to use by default
 
 const stream = function (streamId: string): Stream {
-  const client = communityHelper.getStreamrClient();
+  const client = userHelper.getStreamrClient();
   // Wrap event generation and production into this method
   function produceNewEvent(msg: Message) {
     // Produce the event to the Stream
